# reference-typescript/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Subscription Tracker API
  description: API for managing email-based subscription detection and tracking
  version: 1.0.0
  contact:
    name: Subscription Tracker
    url: https://github.com/your-repo/subscription-tracker
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
security:
  - ApiKeyAuth: []
paths:
...
  /connections/gmail/callback:
    get:
      summary: Handle Gmail OAuth callback
      description: Processes OAuth callback and stores connection
      tags:
        - Gmail Connection
      security: []  # No API key required for OAuth callback
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Google
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: CSRF protection state parameter
      responses:
        '302':
          description: Redirect to success/error page
        '400':
          description: Invalid callback parameters
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (can also be passed as query parameter 'api_key')
...